refactor(ui): split topology layout ownership
