+++ crates/mesh-llm-ui/src/features/dashboard/components/topology/render/useTopologyCanvas.test.tsx
@@ -1,585 +1,42 @@
 // @vitest-environment jsdom
 
-import '@testing-library/jest-dom/vitest'
 import { cleanup, render } from '@testing-library/react'
-import { type MutableRefObject } from 'react'
-import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
+import { afterEach, beforeAll, beforeEach, describe, expect, it } from 'vitest'
 
 import { ENTRY_ANIMATION_DURATION_MS, LINE_REVEAL_DURATION_MS } from '@/features/dashboard/components/topology/helpers'
-import { RENDER_VARIANTS } from '@/features/dashboard/components/topology/theme/render-variants'
 import type {
   EntryAnimation,
   ExitAnimation,
   LineTransition,
   LineRevealAnimation,
   PendingLineTransition,
-  RenderNode,
-  RenderVariant,
-  ScreenNode,
   UpdateTwinkle
 } from '@/features/dashboard/components/topology/types'
-import { buildLineMesh } from '@/features/dashboard/components/topology/render/line-mesh'
-import { useTopologyCanvas } from '@/features/dashboard/components/topology/render/useTopologyCanvas'
+import {
+  Harness,
+  clock,
+  createMutableRef,
+  createRenderNode,
+  createScreenNode,
+  flushAnimationFrame,
+  installTopologyCanvasEnvironment,
+  mocks,
+  resetTopologyCanvasEnvironment
+} from './useTopologyCanvas.test-harness'
 
-type ScenarioName =
-  'join' | 'join-route-change' | 'join-spillover' | 'join-spillover-incoming-only' | 'leave' | 'leave-spillover'
-
-const mocks = vi.hoisted(() => {
-  const state = { scenario: 'join' as ScenarioName }
-
-  const buildProximityLinesMock = vi.fn(
-    (input: {
-      screenNodes: Array<{
-        id: string
-        size: number
-        hitSize: number
-        lineRevealProgress: number
-      }>
-      highlightedNodeIds: Set<string>
-      visiblePairKeys?: Set<string>
-      pairAlphaOverrides?: Map<string, number>
-    }) => {
-      const ids = new Set(input.screenNodes.map((node) => node.id))
-
-      let pairKeys: string[]
-      let pairRouteSignatures: Map<string, string>
-      if (input.visiblePairKeys) {
-        pairKeys = [...input.visiblePairKeys]
-      } else if (state.scenario === 'join') {
-        pairKeys = ids.has('new-node') ? ['anchor::new-node', 'new-node::peer'] : ['anchor::peer', 'stable-a::stable-b']
-      } else if (state.scenario === 'join-route-change') {
-        pairKeys = ['anchor::peer', 'stable-a::stable-b']
-      } else if (state.scenario === 'join-spillover') {
-        pairKeys = ids.has('new-node')
-          ? ['anchor::new-node', 'new-node::peer', 'peer::worker-2', 'worker::remote-2', 'stable-a::stable-b']
-          : ['anchor::peer', 'peer::worker', 'worker::remote', 'stable-a::stable-b']
-      } else if (state.scenario === 'join-spillover-incoming-only') {
-        pairKeys = ids.has('new-node')
-          ? ['anchor::new-node', 'new-node::peer', 'worker::remote-2', 'peer::worker-2', 'stable-a::stable-b']
-          : ['stable-a::stable-b', 'worker::remote', 'peer::worker']
-      } else if (state.scenario === 'leave') {
-        pairKeys = ids.has('replacement') ? ['peer::replacement'] : ['peer::removed-node', 'stable-a::stable-b']
-      } else {
-        pairKeys = ids.has('replacement')
-          ? ['peer::replacement', 'replacement::worker-2', 'worker-2::remote-2', 'stable-a::stable-b']
-          : ['peer::removed-node', 'stable-a::stable-b']
-      }
-
-      // eslint-disable-next-line prefer-const -- assigned after conditional control flow branches determine pairKeys
-      pairRouteSignatures = new Map(
-        pairKeys.map((pairKey) => {
-          if (state.scenario === 'join-route-change' && pairKey === 'anchor::peer') {
-            const signature = ids.has('new-node')
-              ? JSON.stringify({
-                  mode: 'detour',
-                  axis: 'y',
-                  side: 'up',
-                  blockerIds: ['new-node']
-                })
-              : JSON.stringify({ mode: 'straight', blockerIds: [] })
-            return [pairKey, signature]
-          }
-
-          return [pairKey, JSON.stringify({ mode: 'straight', blockerIds: [] })]
-        })
-      )
-
-      return {
-        positions: new Float32Array(),
-        colors: new Float32Array(),
-        pairKeys,
-        pairRouteSignatures
-      }
-    }
-  )
-
-  const createProgramMock = vi.fn(() => ({}))
-  const buildLineFragmentShaderSourceMock = vi.fn(
-    (fragmentSource: string, options: { useStandardDerivatives: boolean }) =>
-      options.useStandardDerivatives ? `derivatives:${fragmentSource}` : fragmentSource
-  )
-
-  return {
-    state,
-    buildProximityLinesMock,
-    createProgramMock,
-    buildLineFragmentShaderSourceMock
-  }
+beforeAll(() => {
+  installTopologyCanvasEnvironment()
 })
 
-vi.mock('@/features/dashboard/components/topology/render/line-builders', () => ({
-  buildProximityLines: mocks.buildProximityLinesMock
-}))
+beforeEach(() => {
+  resetTopologyCanvasEnvironment()
+})
 
-vi.mock('@/features/dashboard/components/topology/render/shaders', () => ({
-  buildLineFragmentShaderSource: mocks.buildLineFragmentShaderSourceMock,
-  createProgram: mocks.createProgramMock,
-  DARK_LINE_FRAGMENT_SHADER: 'dark-line-fragment-shader',
-  DARK_POINT_FRAGMENT_SHADER: 'dark-point-fragment-shader',
-  LIGHT_LINE_FRAGMENT_SHADER: 'light-line-fragment-shader',
-  LIGHT_POINT_FRAGMENT_SHADER: 'light-point-fragment-shader',
-  LINE_VERTEX_SHADER: 'line-vertex-shader',
-  POINT_VERTEX_SHADER: 'point-vertex-shader'
-}))
-
-class MockResizeObserver {
-  observe(): void {}
-  unobserve(): void {}
-  disconnect(): void {}
-}
-
-const clock = { now: 0 }
-const rafState = {
-  nextId: 1,
-  callbacks: new Map<number, FrameRequestCallback>()
-}
-
-let supportsStandardDerivatives = false
-
-let activeGl: ReturnType<typeof createFakeWebGLContext>
-
-function flushAnimationFrame(timestamp = clock.now) {
-  const callbacks = [...rafState.callbacks.values()]
-  rafState.callbacks.clear()
-  for (const callback of callbacks) {
-    callback(timestamp)
-  }
-}
-
-function createFakeWebGLContext() {
-  const attributeLocations = new Map<string, number>([
-    ['a_position', 0],
-    ['a_size', 1],
-    ['a_color', 2],
-    ['a_pulse', 3],
-    ['a_twinkle', 4],
-    ['a_lineCoord', 5]
-  ])
-
-  const uniformLocation = {} as WebGLUniformLocation
-  const buffer = {} as WebGLBuffer
-  const program = {} as WebGLProgram
-
-  return {
-    ARRAY_BUFFER: 0x8892,
-    BLEND: 0x0be2,
-    COLOR_BUFFER_BIT: 0x4000,
-    DYNAMIC_DRAW: 0x88e8,
-    FLOAT: 0x1406,
-    LINES: 0x0001,
-    ONE: 1,
-    ONE_MINUS_SRC_ALPHA: 0x0303,
-    POINTS: 0x0000,
-    SRC_ALPHA: 0x0302,
-    TRIANGLES: 0x0004,
-    blendFunc: vi.fn(),
-    blendFuncSeparate: vi.fn(),
-    bindBuffer: vi.fn(),
-    bufferData: vi.fn(),
-    bufferSubData: vi.fn(),
-    clear: vi.fn(),
-    clearColor: vi.fn(),
-    createBuffer: vi.fn(() => buffer),
-    deleteBuffer: vi.fn(),
-    deleteProgram: vi.fn(),
-    disableVertexAttribArray: vi.fn(),
-    drawArrays: vi.fn(),
-    enable: vi.fn(),
-    enableVertexAttribArray: vi.fn(),
-    getAttribLocation: vi.fn((_program: WebGLProgram, name: string) => attributeLocations.get(name) ?? 0),
-    getExtension: vi.fn((name: string) =>
-      name === 'OES_standard_derivatives' && supportsStandardDerivatives ? {} : null
-    ),
-    getUniformLocation: vi.fn(() => uniformLocation),
-    lineWidth: vi.fn(),
-    uniform1f: vi.fn(),
-    uniform2f: vi.fn(),
-    useProgram: vi.fn(),
-    vertexAttribPointer: vi.fn(),
-    viewport: vi.fn(),
-    createProgram: vi.fn(() => program)
-  }
-}
-
-function createRenderNode(id: string, overrides: Partial<RenderNode> = {}): RenderNode {
-  return {
-    id,
-    label: id,
-    subtitle: id,
-    role: 'Worker',
-    latencyLabel: '0ms',
-    vramLabel: '0 GB',
-    modelLabel: '',
-    gpuLabel: '',
-    x: 0.1,
-    y: 0.1,
-    size: 10,
-    color: [0.2, 0.3, 0.4, 1],
-    lineColor: [0.2, 0.3, 0.4, 1],
-    pulse: 0,
-    selectedModelMatch: false,
-    z: 0,
-    ...overrides
-  }
-}
-
-function createScreenNode(node: RenderNode): ScreenNode {
-  return {
-    ...node,
-    px: node.x * 800,
-    py: node.y * 600,
-    hitSize: node.size,
-    lineRevealProgress: 1
-  }
-}
-
-function createMutableRef<T>(current: T): MutableRefObject<T> {
-  return { current }
-}
-
-function Harness({
-  animationRef,
-  canvasRef,
-  exitAnimationRef,
-  hostRef,
-  hoveredNodeIdRef,
-  lastScreenPositionsRef,
-  lineTransitionRef,
-  lineRevealRef,
-  panRef,
-  pendingLineTransitionRef,
-  renderNodes,
-  renderVariant = RENDER_VARIANTS.light,
-  screenNodesRef,
-  seenNodeIdsRef,
-  selfNodeId,
-  twinkleAnimationRef,
-  zoomRef
-}: {
-  canvasRef: MutableRefObject<HTMLCanvasElement | null>
-  hostRef: MutableRefObject<HTMLDivElement | null>
-  screenNodesRef: MutableRefObject<ScreenNode[]>
-  animationRef: MutableRefObject<Map<string, EntryAnimation>>
-  lineRevealRef: MutableRefObject<Map<string, LineRevealAnimation>>
-  exitAnimationRef: MutableRefObject<Map<string, ExitAnimation>>
-  twinkleAnimationRef: MutableRefObject<Map<string, UpdateTwinkle>>
-  pendingLineTransitionRef: MutableRefObject<PendingLineTransition | null>
-  lineTransitionRef: MutableRefObject<LineTransition | null>
-  lastScreenPositionsRef: MutableRefObject<Map<string, { x: number; y: number }>>
-  seenNodeIdsRef: MutableRefObject<Set<string>>
-  hoveredNodeIdRef: MutableRefObject<string | null>
-  zoomRef: MutableRefObject<number>
-  panRef: MutableRefObject<{ x: number; y: number }>
-  renderNodes: RenderNode[]
-  renderVariant?: RenderVariant
-  selfNodeId?: string
-}) {
-  useTopologyCanvas({
-    animationRef,
-    canvasRef,
-    exitAnimationRef,
-    hostRef,
-    hoveredNodeIdRef,
-    lastScreenPositionsRef,
-    lineTransitionRef,
-    lineRevealRef,
-    panRef,
-    pendingLineTransitionRef,
-    renderNodes,
-    renderVariant,
-    screenNodesRef,
-    seenNodeIdsRef,
-    selfNodeId,
-    twinkleAnimationRef,
-    zoomRef
-  })
-
-  return (
-    <div ref={hostRef}>
-      <canvas ref={canvasRef} />
-    </div>
-  )
-}
+afterEach(() => {
+  cleanup()
+})
 
 describe('useTopologyCanvas light transition regressions', () => {
-  beforeAll(() => {
-    Object.defineProperty(window, 'ResizeObserver', {
-      configurable: true,
-      writable: true,
-      value: MockResizeObserver
-    })
-
-    Object.defineProperty(globalThis, 'WebGLRenderingContext', {
-      configurable: true,
-      writable: true,
-      value: function WebGLRenderingContext() {}
-    })
-
-    Object.defineProperty(window, 'devicePixelRatio', {
-      configurable: true,
-      value: 1
-    })
-
-    Object.defineProperty(HTMLElement.prototype, 'getBoundingClientRect', {
-      configurable: true,
-      value: () => ({
-        width: 800,
-        height: 600,
-        x: 0,
-        y: 0,
-        top: 0,
-        left: 0,
-        right: 800,
-        bottom: 600,
-        toJSON: () => ({})
-      })
-    })
-
-    Object.defineProperty(window, 'requestAnimationFrame', {
-      configurable: true,
-      writable: true,
-      value: (callback: FrameRequestCallback) => {
-        const id = rafState.nextId
-        rafState.nextId += 1
-        rafState.callbacks.set(id, callback)
-        return id
-      }
-    })
-
-    Object.defineProperty(window, 'cancelAnimationFrame', {
-      configurable: true,
-      writable: true,
-      value: (id: number) => {
-        rafState.callbacks.delete(id)
-      }
-    })
-
-    Object.defineProperty(performance, 'now', {
-      configurable: true,
-      value: () => clock.now
-    })
-  })
-
-  beforeEach(() => {
-    mocks.state.scenario = 'join'
-    mocks.buildProximityLinesMock.mockClear()
-    mocks.buildLineFragmentShaderSourceMock.mockClear()
-    mocks.createProgramMock.mockClear()
-    rafState.callbacks.clear()
-    rafState.nextId = 1
-    clock.now = 0
-    supportsStandardDerivatives = false
-
-    activeGl = createFakeWebGLContext()
-    Object.defineProperty(HTMLCanvasElement.prototype, 'getContext', {
-      configurable: true,
-      value: (contextType: string) => (contextType === 'webgl' ? activeGl : null)
-    })
-  })
-
-  afterEach(() => {
-    cleanup()
-  })
-
-  it('uses the render variant line shader and applies line blend state separately', () => {
-    const applyLineBlendMode = vi.fn()
-    const applyPointBlendMode = vi.fn()
-    const renderVariant: RenderVariant = {
-      ...RENDER_VARIANTS.dark,
-      lineFragmentShader: 'custom-line-fragment-shader',
-      pointFragmentShader: 'custom-point-fragment-shader',
-      applyLineBlendMode,
-      applyPointBlendMode,
-      buildLines: () => ({
-        positions: new Float32Array([0, 0, 80, 80]),
-        colors: new Float32Array([0.4, 0.6, 0.9, 0.28, 0.4, 0.6, 0.9, 0.12])
-      })
-    }
-
-    render(
-      <Harness
-        animationRef={createMutableRef<Map<string, EntryAnimation>>(new Map())}
-        canvasRef={createMutableRef<HTMLCanvasElement | null>(null)}
-        exitAnimationRef={createMutableRef<Map<string, ExitAnimation>>(new Map())}
-        hostRef={createMutableRef<HTMLDivElement | null>(null)}
-        hoveredNodeIdRef={createMutableRef<string | null>(null)}
-        lastScreenPositionsRef={createMutableRef<Map<string, { x: number; y: number }>>(new Map())}
-        lineTransitionRef={createMutableRef<LineTransition | null>(null)}
-        lineRevealRef={createMutableRef<Map<string, LineRevealAnimation>>(new Map())}
-        panRef={createMutableRef({ x: 0, y: 0 })}
-        pendingLineTransitionRef={createMutableRef<PendingLineTransition | null>(null)}
-        renderNodes={[createRenderNode('peer', { x: 0.2, y: 0.3 })]}
-        renderVariant={renderVariant}
-        screenNodesRef={createMutableRef<ScreenNode[]>([])}
-        seenNodeIdsRef={createMutableRef<Set<string>>(new Set(['peer']))}
-        selfNodeId="peer"
-        twinkleAnimationRef={createMutableRef<Map<string, UpdateTwinkle>>(new Map())}
-        zoomRef={createMutableRef(1)}
-      />
-    )
-
-    expect(mocks.createProgramMock).toHaveBeenNthCalledWith(
-      1,
-      expect.anything(),
-      'point-vertex-shader',
-      'custom-point-fragment-shader'
-    )
-    expect(mocks.buildLineFragmentShaderSourceMock).toHaveBeenCalledWith('custom-line-fragment-shader', {
-      useStandardDerivatives: false
-    })
-    expect(mocks.createProgramMock).toHaveBeenNthCalledWith(
-      2,
-      expect.anything(),
-      'line-vertex-shader',
-      'custom-line-fragment-shader'
-    )
-    const expectedMesh = buildLineMesh({
-      positions: new Float32Array([0, 0, 80, 80]),
-      colors: new Float32Array([0.4, 0.6, 0.9, 0.28, 0.4, 0.6, 0.9, 0.12]),
-      lineWidthPx: renderVariant.lineWidthPx,
-      devicePixelRatio: 1
-    })
-    expect(applyLineBlendMode).toHaveBeenCalledWith(activeGl)
-    expect(applyPointBlendMode).toHaveBeenCalledWith(activeGl)
-    expect(activeGl.drawArrays).toHaveBeenNthCalledWith(1, activeGl.TRIANGLES, 0, expectedMesh.positions.length / 2)
-    expect(activeGl.drawArrays).toHaveBeenNthCalledWith(2, activeGl.POINTS, 0, 1)
-    expect(activeGl.lineWidth).not.toHaveBeenCalled()
-    expect(applyLineBlendMode.mock.invocationCallOrder[0]).toBeLessThan(activeGl.drawArrays.mock.invocationCallOrder[0])
-    expect(applyPointBlendMode.mock.invocationCallOrder[0]).toBeGreaterThan(
-      applyLineBlendMode.mock.invocationCallOrder[0]
-    )
-    expect(applyPointBlendMode.mock.invocationCallOrder[0]).toBeLessThan(
-      activeGl.drawArrays.mock.invocationCallOrder[1]
-    )
-  })
-
-  it('enables derivative-aware line shader source when the WebGL extension is available', () => {
-    supportsStandardDerivatives = true
-    activeGl = createFakeWebGLContext()
-    Object.defineProperty(HTMLCanvasElement.prototype, 'getContext', {
-      configurable: true,
-      value: (contextType: string) => (contextType === 'webgl' ? activeGl : null)
-    })
-
-    render(
-      <Harness
-        animationRef={createMutableRef<Map<string, EntryAnimation>>(new Map())}
-        canvasRef={createMutableRef<HTMLCanvasElement | null>(null)}
-        exitAnimationRef={createMutableRef<Map<string, ExitAnimation>>(new Map())}
-        hostRef={createMutableRef<HTMLDivElement | null>(null)}
-        hoveredNodeIdRef={createMutableRef<string | null>(null)}
-        lastScreenPositionsRef={createMutableRef<Map<string, { x: number; y: number }>>(new Map())}
-        lineTransitionRef={createMutableRef<LineTransition | null>(null)}
-        lineRevealRef={createMutableRef<Map<string, LineRevealAnimation>>(new Map())}
-        panRef={createMutableRef({ x: 0, y: 0 })}
-        pendingLineTransitionRef={createMutableRef<PendingLineTransition | null>(null)}
-        renderNodes={[createRenderNode('peer', { x: 0.2, y: 0.3 })]}
-        screenNodesRef={createMutableRef<ScreenNode[]>([])}
-        seenNodeIdsRef={createMutableRef<Set<string>>(new Set(['peer']))}
-        selfNodeId="peer"
-        twinkleAnimationRef={createMutableRef<Map<string, UpdateTwinkle>>(new Map())}
-        zoomRef={createMutableRef(1)}
-      />
-    )
-
-    expect(mocks.buildLineFragmentShaderSourceMock).toHaveBeenCalledWith(RENDER_VARIANTS.light.lineFragmentShader, {
-      useStandardDerivatives: true
-    })
-    expect(mocks.createProgramMock).toHaveBeenNthCalledWith(
-      2,
-      expect.anything(),
-      'line-vertex-shader',
-      `derivatives:${RENDER_VARIANTS.light.lineFragmentShader}`
-    )
-  })
-
-  it('keeps light line rendering unchanged when hover is active', () => {
-    render(
-      <Harness
-        animationRef={createMutableRef<Map<string, EntryAnimation>>(new Map())}
-        canvasRef={createMutableRef<HTMLCanvasElement | null>(null)}
-        exitAnimationRef={createMutableRef<Map<string, ExitAnimation>>(new Map())}
-        hostRef={createMutableRef<HTMLDivElement | null>(null)}
-        hoveredNodeIdRef={createMutableRef<string | null>('peer')}
-        lastScreenPositionsRef={createMutableRef<Map<string, { x: number; y: number }>>(new Map())}
-        lineTransitionRef={createMutableRef<LineTransition | null>(null)}
-        lineRevealRef={createMutableRef<Map<string, LineRevealAnimation>>(new Map())}
-        panRef={createMutableRef({ x: 0, y: 0 })}
-        pendingLineTransitionRef={createMutableRef<PendingLineTransition | null>(null)}
-        renderNodes={[
-          createRenderNode('anchor', { x: 0.2, y: 0.2 }),
-          createRenderNode('peer', { x: 0.8, y: 0.2 }),
-          createRenderNode('stable-a', { x: 0.2, y: 0.8 }),
-          createRenderNode('stable-b', { x: 0.8, y: 0.8 })
-        ]}
-        screenNodesRef={createMutableRef<ScreenNode[]>([])}
-        seenNodeIdsRef={createMutableRef<Set<string>>(new Set(['anchor', 'peer', 'stable-a', 'stable-b']))}
-        selfNodeId="peer"
-        twinkleAnimationRef={createMutableRef<Map<string, UpdateTwinkle>>(new Map())}
-        zoomRef={createMutableRef(1)}
-      />
-    )
-
-    expect(mocks.buildProximityLinesMock).toHaveBeenCalledTimes(1)
-    const baseCall = mocks.buildProximityLinesMock.mock.calls[0]?.[0]
-
-    expect([...baseCall.highlightedNodeIds]).toEqual([])
-    expect(baseCall.visiblePairKeys).toBeUndefined()
-    expect(new Set(baseCall.screenNodes.map((node: { id: string }) => node.id))).toEqual(
-      new Set(['anchor', 'peer', 'stable-a', 'stable-b'])
-    )
-    expect(activeGl.drawArrays).toHaveBeenCalledTimes(1)
-    expect(activeGl.drawArrays).toHaveBeenCalledWith(activeGl.POINTS, 0, 4)
-  })
-
-  it('does not change line-builder screen-node sizes when hover changes in the normal light path', () => {
-    const hoveredNodeIdRef = createMutableRef<string | null>(null)
-    const renderNodes = [
-      createRenderNode('anchor', { x: 0.2, y: 0.2, size: 16 }),
-      createRenderNode('peer', { x: 0.8, y: 0.2, size: 14 }),
-      createRenderNode('stable-a', { x: 0.2, y: 0.8, size: 10 }),
-      createRenderNode('stable-b', { x: 0.8, y: 0.8, size: 11 })
-    ]
-    const refs = {
-      animationRef: createMutableRef<Map<string, EntryAnimation>>(new Map()),
-      canvasRef: createMutableRef<HTMLCanvasElement | null>(null),
-      exitAnimationRef: createMutableRef<Map<string, ExitAnimation>>(new Map()),
-      hostRef: createMutableRef<HTMLDivElement | null>(null),
-      hoveredNodeIdRef,
-      lastScreenPositionsRef: createMutableRef<Map<string, { x: number; y: number }>>(new Map()),
-      lineTransitionRef: createMutableRef<LineTransition | null>(null),
-      lineRevealRef: createMutableRef<Map<string, LineRevealAnimation>>(new Map()),
-      panRef: createMutableRef({ x: 0, y: 0 }),
-      pendingLineTransitionRef: createMutableRef<PendingLineTransition | null>(null),
-      screenNodesRef: createMutableRef<ScreenNode[]>([]),
-      seenNodeIdsRef: createMutableRef<Set<string>>(new Set(renderNodes.map((node) => node.id))),
-      twinkleAnimationRef: createMutableRef<Map<string, UpdateTwinkle>>(new Map()),
-      zoomRef: createMutableRef(1)
-    }
-
-    const { rerender } = render(<Harness {...refs} renderNodes={renderNodes} selfNodeId="peer" />)
-
-    const baseCall = mocks.buildProximityLinesMock.mock.calls[0]?.[0]
-    const baseIds = baseCall.screenNodes.map((node: { id: string }) => node.id)
-    const baseSizes = new Map(baseCall.screenNodes.map((node: { id: string; size: number }) => [node.id, node.size]))
-    expect(baseIds).toEqual(['anchor', 'peer', 'stable-b', 'stable-a'])
-
-    mocks.buildProximityLinesMock.mockClear()
-    hoveredNodeIdRef.current = 'peer'
-    rerender(<Harness {...refs} renderNodes={renderNodes.map((node) => ({ ...node }))} selfNodeId="peer" />)
-    flushAnimationFrame()
-
-    expect(mocks.buildProximityLinesMock).not.toHaveBeenCalled()
-    expect(refs.screenNodesRef.current.map((node) => node.id)).toEqual(['peer', 'anchor', 'stable-b', 'stable-a'])
-    const hoveredScreenNode = refs.screenNodesRef.current.find((node) => node.id === 'peer')
-    expect(hoveredScreenNode?.hitSize).toBeGreaterThan(hoveredScreenNode?.size ?? 0)
-    const stableLineNodes = [...refs.screenNodesRef.current]
-      .sort((left, right) => left.id.localeCompare(right.id))
-      .map((node) => node.id)
-    expect(stableLineNodes.slice().sort()).toEqual(baseIds.slice().sort())
-    expect(new Map(refs.screenNodesRef.current.map((node) => [node.id, node.size]))).toEqual(baseSizes)
-  })
-
   it('keeps snapshot lines stable during the join pre-reveal branch', () => {
     mocks.state.scenario = 'join'
